requirements.txt.txt
python-binance==1.0.19
openpyxl==3.1.2
